Update `locale-router-static` example
