+++ examples/locale-router-static/src/lib/translations/index.js
@@ -1,3 +1,4 @@
+import { dev } from '$app/environment';
 import i18n from 'sveltekit-i18n';
 import en from './en';
 import de from './de';
@@ -6,6 +7,9 @@ import lang from './lang';
 
 /** @type {import('sveltekit-i18n').Config} */
 const config = {
+  log: {
+    level: dev ? 'warn' : 'error',
+  },
   translations: {
     en: {
       ...en,
@@ -24,4 +28,4 @@ const config = {
 
 export const defaultLocale = 'en';
 
-export const { t, locale, locales, loading, loadTranslations, translations } = new i18n(config);
+export const { t, locale, locales, loading, setLocale, setRoute, translations } = new i18n(config);
