+++ examples/locale-router-static/src/routes/+error.svelte
@@ -1,16 +1,3 @@
-<script context="module">
-  import { loadTranslations } from '$lib/translations';
-  
-  /** @type {import('@sveltejs/kit').Load} */
-  export const load = async ({ stuff, props }) => {
-
-    await loadTranslations(stuff.lang, 'error');
-
-    return {};
-  }
-</script>
-
-
 <script>
   import { t, locale } from '$lib/translations';
   import { page } from '$app/stores';
