+++ examples/locale-router-static/src/routes/[lang]/+page.svelte
@@ -1,7 +1,7 @@
 <script>
-  import { t } from '$lib/translations'; 
+  import { t } from '$lib/translations';
 
-  const link = 'https://kit.svelte.dev'
+  const link = 'https://kit.svelte.dev';
 </script>
 
 <h1>{$t('home.title')}</h1>
