+++ examples/locale-router-static/src/routes/[lang]/[...error]/+page@.svelte
@@ -1,0 +1,7 @@
+<script>
+  import Error from '../../+error.svelte';
+
+  export let data;
+</script>
+
+<Error status={data.status} />
