+++ examples/locale-router-static/svelte.config.js
@@ -10,7 +10,7 @@ const config = {
   kit: {
     adapter: adapter(),
     prerender: {
-      default: true,
+      // default: true,
       // NOTE: You can modify your exported error pages here.
       entries: supportedLocales.reduce((acc, locale) => [...acc, `/${locale}`, `/${locale}/401`, `/${locale}/403`, `/${locale}/404`, `/${locale}/500`], ['*']),
     },
